+++ ruoyi-admin/src/main/java/com/ruoyi/system/service/impl/StoragequitbillServiceImpl.java
@@ -114,13 +114,13 @@ public class StoragequitbillServiceImpl implements IStoragequitbillService
             Storage storage=new Storage();
             WarehouseRecord warehouseRecord=new WarehouseRecord();
             storage.setStocks(storagequitdetail.getCounts());
-            storage.setMoney(new BigDecimal(storagequitdetail.getMoney()).setScale(2,BigDecimal.ROUND_HALF_UP));
+            storage.setMoney(new BigDecimal(storagequitdetail.getMoney()).setScale(2,BigDecimal.ROUND_HALF_UP).doubleValue());
             storage.setId(storagequitdetail.getSid());
             storageMapper.updatereducestocks(storage);
             if(storageMapper.selectStorageById(storage.getId()).getStocks()==0){
                 Storage storage1=new Storage();
                 storage1.setId(storage.getId());
-                storage1.setMoney(new BigDecimal(0));
+                storage1.setMoney(new BigDecimal(0).doubleValue());
                 storageMapper.updateStorage(storage1);
             }
             warehouseRecord.setType(Const.WarehouseRecordStatus.STORAGE_QUIT_HC);
